[utils] dep-of-std feature and moved from heapless dependency to arrayvec
