+++ utils/Cargo.toml
@@ -4,7 +4,16 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-heapless = "0.8.0"
 macros = { path = "../macros" }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 int-enum = "1.1"
+
+arrayvec = { version = "0.7.6", default-features = false }
+
+
+core = { version = "1.0.0", optional = true, package = "rustc-std-workspace-core" }
+alloc = { version = "1.0.0", optional = true, package = "rustc-std-workspace-alloc" }
+compiler_builtins = { version = "0.1", optional = true }
+[features]
+default = []
+rustc-dep-of-std = ["core", "alloc", "compiler_builtins/rustc-dep-of-std"]
